Per STR #230, changed call to fabsf() to the standard C library call fabs(3M). On some platforms (e.g. Solaris SPARC), single precision math functions are not declared in the standard headers nor implemented in the standard libraries. There does not seem to be any advantage to the single precision version here since the argument ultimately passed to handle_drag() is a double.
git-svn-id: file:///fltk/svn/trunk@3467 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/Fl_Scrollbar.cxx
@@ -1,5 +1,5 @@
 //
-// "$Id: Fl_Scrollbar.cxx,v 1.70 2004/06/11 08:07:18 spitzak Exp $"
+// "$Id: Fl_Scrollbar.cxx,v 1.71 2004/06/24 21:22:30 xpxqx Exp $"
 //
 // Scroll bar widget for the Fast Light Tool Kit (FLTK).
 //
@@ -220,9 +220,9 @@ int Scrollbar::handle(int event) {
     which_pushed = NOTHING;
     return 1;
   case MOUSEWHEEL: {
-    float n = (vertical() ? -event_dy() : event_dx())
+    double n = (vertical() ? -event_dy() : event_dx())
       * style()->wheel_scroll_lines() * linesize();
-    if (fabsf(n) > pagesize()) n = (n<0)?-pagesize():pagesize();
+    if (fabs(n) > pagesize()) n = (n<0)?-pagesize():pagesize();
     handle_drag(value()+n);
     return 1;
   }
@@ -313,5 +313,5 @@ Scrollbar::Scrollbar(int X, int Y, int W, int H, const char* L)
 }
 
 //
-// End of "$Id: Fl_Scrollbar.cxx,v 1.70 2004/06/11 08:07:18 spitzak Exp $".
+// End of "$Id: Fl_Scrollbar.cxx,v 1.71 2004/06/24 21:22:30 xpxqx Exp $".
 //
